requirements.txt
streamlit==1.45.1
PyMuPDF==1.25.5
pillow==11.2.1